services/web/requirements.txt
Flask==2.3.2
Flask[async]
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
bleach==4.1.0
